refactor: Split comparators.py into dominance / nds / comparators
