+++ src/saealib/comparators.py
@@ -8,309 +8,21 @@ for ranking and comparing solutions in single- and multi-objective optimization.
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Protocol
+from typing import TYPE_CHECKING
 
 import numpy as np
 
 from saealib._deprecated import deprecated_param, warn_deprecated
+from saealib.dominance import Dominator, EpsilonDominator, ParetoDominator
+from saealib.nds import (
+    NonDominatedSorter,
+    crowding_distance_all_fronts,
+    non_dominated_sort,
+    spea2_fitness,
+)
 
 if TYPE_CHECKING:
     from saealib.population import Population
-
-
-# ---------------------------------------------------------------------------
-# Dominator abstraction
-# ---------------------------------------------------------------------------
-
-
-class Dominator(ABC):
-    """
-    Abstract base for dominance predicates.
-
-    A ``Dominator`` encapsulates the definition of dominance between
-    objective vectors independently of the sorting algorithm.  Two concrete
-    operations are required and MUST agree with each other:
-
-    - ``dominance_matrix`` — batched NxN boolean matrix (primary operation).
-    - ``dominates`` — scalar pairwise predicate, derived from the matrix to
-      guarantee consistency.
-
-    Parameters are the same for both methods:
-
-    Parameters
-    ----------
-    f : np.ndarray
-        Objective matrix. shape: (n, n_obj)  [``dominance_matrix``]
-    fa, fb : np.ndarray
-        Objective vectors. shape: (n_obj,)  [``dominates``]
-    direction : np.ndarray or None
-        Per-objective optimization direction: +1 = maximize, -1 = minimize.
-        ``None`` defaults to minimization for all objectives.
-    """
-
-    @abstractmethod
-    def dominance_matrix(
-        self,
-        f: np.ndarray,
-        direction: np.ndarray | None = None,
-    ) -> np.ndarray:
-        """
-        Compute the NxN boolean dominance matrix.
-
-        ``D[i, j]`` is True if row i dominates row j.
-
-        Parameters
-        ----------
-        f : np.ndarray
-            Objective matrix. shape: (n, n_obj).  Must contain no NaN values
-            (callers are responsible for pre-filtering).
-        direction : np.ndarray or None
-            Per-objective direction. ``None`` → minimize all.
-
-        Returns
-        -------
-        np.ndarray
-            Boolean matrix of shape (n, n).
-        """
-
-    def dominates(
-        self,
-        fa: np.ndarray,
-        fb: np.ndarray,
-        direction: np.ndarray | None = None,
-    ) -> bool:
-        """
-        Return True if fa dominates fb.
-
-        Derived from ``dominance_matrix`` on a 2-row stack to guarantee
-        agreement with the batched path.  NaN values in fa always return
-        False (consistent with ``dominance_matrix`` assuming finite input).
-
-        Parameters
-        ----------
-        fa, fb : np.ndarray
-            Objective vectors. shape: (n_obj,)
-        direction : np.ndarray or None
-            Per-objective direction. ``None`` → minimize all.
-
-        Returns
-        -------
-        bool
-        """
-        fa = np.asarray(fa, dtype=float)
-        fb = np.asarray(fb, dtype=float)
-        if np.any(np.isnan(fa)):
-            return False
-        # Replace NaN in fb with +inf so it appears infinitely bad (dominated).
-        fb_safe = np.where(np.isnan(fb), np.inf, fb)
-        stacked = np.stack([fa, fb_safe])
-        return bool(self.dominance_matrix(stacked, direction)[0, 1])
-
-
-class ParetoDominator(Dominator):
-    """
-    Pareto dominance predicate.
-
-    ``fa`` dominates ``fb`` iff ``fa`` is at most as large as ``fb`` in all
-    objectives and strictly smaller in at least one (after applying the
-    direction transform).
-
-    This is the default dominance relation used throughout the library.
-    """
-
-    def dominance_matrix(
-        self,
-        f: np.ndarray,
-        direction: np.ndarray | None = None,
-    ) -> np.ndarray:
-        """
-        Compute the NxN Pareto-dominance matrix.
-
-        Applies the ``f * (-direction)`` transform so that smaller is always
-        better, then accumulates per-objective ``leq_all`` / ``less_any``
-        comparisons.  Peak memory is O(N²); no (N, N, M) tensor is created.
-
-        Parameters
-        ----------
-        f : np.ndarray
-            Objective matrix. shape: (n, n_obj).  Must contain no NaN values.
-        direction : np.ndarray or None
-            Per-objective direction. ``None`` → minimize all.
-
-        Returns
-        -------
-        np.ndarray
-            Boolean matrix of shape (n, n).  ``D[i, j]`` = True iff row i
-            Pareto-dominates row j.
-        """
-        g = np.asarray(f, dtype=float)
-        if direction is not None:
-            g = g * (-direction)
-        k, m = g.shape
-        leq_all = np.ones((k, k), dtype=bool)
-        less_any = np.zeros((k, k), dtype=bool)
-        for obj in range(m):
-            col = g[:, obj]
-            leq_all &= col[:, None] <= col[None, :]
-            less_any |= col[:, None] < col[None, :]
-        return leq_all & less_any
-
-
-class EpsilonDominator(Dominator):
-    """
-    ε-dominance predicate (Laumanns et al., 2002).
-
-    Quantizes each objective into ε-boxes and applies ordinary Pareto
-    dominance on the resulting integer box coordinates.  Two solutions that
-    fall in the same ε-box are mutually non-dominating; a solution in a
-    strictly better box (lower box index in every minimization objective,
-    higher in every maximization objective) dominates the other.
-
-    Two quantization modes are supported:
-
-    - **additive** (default): box index = ``floor(f_i / eps_i)``.
-      As ``eps → 0`` the boxes shrink to individual points and the relation
-      recovers ordinary Pareto dominance.
-
-    - **multiplicative**: box index = ``floor(log f_i / log(1 + eps_i))``.
-      Requires strictly positive objective values (``f > 0``).
-
-    Parameters
-    ----------
-    eps : float or np.ndarray
-        Box width per objective.  A scalar applies the same width to every
-        objective; an array of shape ``(n_obj,)`` sets per-objective widths.
-        All values must be strictly positive (> 0).
-    mode : {"additive", "multiplicative"}
-        Quantization mode.  Default is ``"additive"``.
-
-    Raises
-    ------
-    ValueError
-        If any element of ``eps`` is not strictly positive (≤ 0).
-    ValueError
-        If ``mode`` is not one of ``"additive"`` or ``"multiplicative"``.
-
-    Notes
-    -----
-    Additive box rule::
-
-        b_i = floor(f_i / eps_i)
-
-    Multiplicative box rule (requires f_i > 0 for all i)::
-
-        b_i = floor(log(f_i) / log(1 + eps_i))
-
-    Because ``log(1 + eps_i) = log1p(eps_i)`` is used internally, the
-    multiplicative rule is numerically stable for small ``eps``.
-
-    The dominance-matrix computation delegates entirely to an internal
-    :class:`ParetoDominator` instance operating on the quantized box
-    coordinates.  This guarantees that :meth:`dominates` (inherited from
-    :class:`Dominator`) and :meth:`dominance_matrix` always agree.
-
-    References
-    ----------
-    .. [1] Laumanns, M., Thiele, L., Deb, K., & Zitzler, E. (2002).
-       Combining Convergence and Diversity in Evolutionary Multiobjective
-       Optimization. *Evolutionary Computation*, 10(3), 263-282.
-       https://doi.org/10.1162/106365602760234108
-    """
-
-    def __init__(
-        self,
-        eps: float | np.ndarray,
-        mode: str = "additive",
-    ) -> None:
-        eps_arr = np.atleast_1d(np.asarray(eps, dtype=float))
-        if np.any(eps_arr <= 0):
-            raise ValueError(
-                f"All eps values must be strictly positive (> 0); got {eps_arr}"
-            )
-        if mode not in ("additive", "multiplicative"):
-            raise ValueError(
-                f"mode must be 'additive' or 'multiplicative'; got {mode!r}"
-            )
-        self._eps = eps_arr
-        self._mode = mode
-        # Internal delegate: Pareto dominance on box coordinates.
-        self._pareto = ParetoDominator()
-
-    @property
-    def eps(self) -> np.ndarray:
-        """Box widths per objective."""
-        return self._eps
-
-    @property
-    def mode(self) -> str:
-        """Quantization mode ('additive' or 'multiplicative')."""
-        return self._mode
-
-    def _quantize(self, f: np.ndarray) -> np.ndarray:
-        """
-        Map objective matrix *f* to integer ε-box coordinates.
-
-        Parameters
-        ----------
-        f : np.ndarray
-            Objective matrix, shape (n, n_obj).
-
-        Returns
-        -------
-        np.ndarray
-            Integer box-index matrix, shape (n, n_obj), dtype float64
-            (floor of real-valued box coordinates; kept as float so that
-            ParetoDominator can operate on them without type issues).
-
-        Raises
-        ------
-        ValueError
-            In multiplicative mode, if any element of *f* is ≤ 0.
-        """
-        if self._mode == "additive":
-            return np.floor(f / self._eps)
-        # multiplicative mode
-        if np.any(f <= 0):
-            raise ValueError(
-                "Multiplicative ε-dominance requires all objective values to be "
-                "strictly positive (f > 0).  Found non-positive values in f."
-            )
-        return np.floor(np.log(f) / np.log1p(self._eps))
-
-    def dominance_matrix(
-        self,
-        f: np.ndarray,
-        direction: np.ndarray | None = None,
-    ) -> np.ndarray:
-        """
-        Compute the NxN ε-dominance matrix.
-
-        Quantizes *f* into ε-box coordinates and delegates to
-        :class:`ParetoDominator` on those coordinates so that direction
-        handling and the accumulation logic are not duplicated.
-
-        ``D[i, j]`` is True if row i ε-dominates row j.
-
-        Parameters
-        ----------
-        f : np.ndarray
-            Objective matrix. shape: (n, n_obj).  Must contain no NaN values
-            (callers are responsible for pre-filtering).
-        direction : np.ndarray or None
-            Per-objective direction. ``None`` → minimize all.
-
-        Returns
-        -------
-        np.ndarray
-            Boolean matrix of shape (n, n).
-
-        Raises
-        ------
-        ValueError
-            In multiplicative mode, if any element of *f* is ≤ 0.
-        """
-        b = self._quantize(np.asarray(f, dtype=float))
-        return self._pareto.dominance_matrix(b, direction)
 
 
 class Comparator(ABC):
@@ -583,455 +295,6 @@ class WeightedSumComparator(Comparator):
         elif sa < sb - self.eps_obj:
             return 1
         return 0
-
-
-# ---------------------------------------------------------------------------
-# Non-dominated sorting utilities
-# ---------------------------------------------------------------------------
-
-# Module-level singleton used by legacy wrappers (_pareto_dominates,
-# _dominance_matrix) so callers that import them directly stay correct.
-_PARETO_DOMINATOR: ParetoDominator = ParetoDominator()
-
-
-def _pareto_dominates(
-    fa: np.ndarray, fb: np.ndarray, direction: np.ndarray | None = None
-) -> bool:
-    """
-    Return True if fa Pareto-dominates fb.
-
-    NaN values in fa are treated as non-dominating (returns False).
-
-    .. deprecated::
-        Thin wrapper kept for backward compatibility.  Use
-        ``ParetoDominator().dominates(fa, fb, direction)`` instead.
-
-    Parameters
-    ----------
-    fa, fb : np.ndarray
-        Objective vectors to compare.
-    direction : np.ndarray or None
-        Per-objective optimization direction: +1 = maximize, -1 = minimize.
-        None defaults to minimization for all objectives.
-    """
-    return _PARETO_DOMINATOR.dominates(fa, fb, direction)
-
-
-def _dominance_matrix(g: np.ndarray) -> np.ndarray:
-    """
-    Compute the NxN boolean dominance matrix from a direction-transformed matrix.
-
-    ``D[i, j]`` is True if row i Pareto-dominates row j.  The input ``g``
-    must already have the direction transform applied (smaller = better for
-    every objective) and must contain no NaN values.
-
-    .. deprecated::
-        Thin wrapper kept for backward compatibility.  Use
-        ``ParetoDominator().dominance_matrix(g)`` instead.
-
-    Parameters
-    ----------
-    g : np.ndarray
-        Objective matrix after direction transform. shape: (k, n_obj).
-        Must contain no NaN values.
-
-    Returns
-    -------
-    np.ndarray
-        Boolean matrix of shape (k, k).
-    """
-    # g is already direction-transformed; pass direction=None so no second
-    # transform is applied inside ParetoDominator.
-    return _PARETO_DOMINATOR.dominance_matrix(g)
-
-
-def non_dominated_sort(
-    f: np.ndarray,
-    direction: np.ndarray | None = None,
-    *,
-    dominator: Dominator | None = None,
-) -> tuple[np.ndarray, list[list[int]]]:
-    """
-    Non-dominated sorting (Deb et al., 2002) via a vectorized dominance matrix.
-
-    Time is O(MN^2), but the pairwise dominance relations are computed with a
-    NumPy dominance matrix accumulated one objective at a time (peak memory
-    O(N^2), no (N, N, M) tensor), replacing the original Python double loop.
-
-    NaN rows are treated as infinitely bad and placed in the last front.
-
-    Parameters
-    ----------
-    f : np.ndarray
-        Objective matrix. shape: (n, n_obj)
-    direction : np.ndarray or None
-        Per-objective optimization direction: +1 = maximize, -1 = minimize.
-        None defaults to minimization for all objectives.
-    dominator : Dominator or None
-        Dominance predicate to use.  ``None`` defaults to
-        ``ParetoDominator`` (standard Pareto dominance).
-
-    Returns
-    -------
-    ranks : np.ndarray shape (n,)
-        Pareto front index for each individual (0 = first/best front).
-    fronts : list[list[int]]
-        fronts[i] contains the local indices of individuals in front i.
-    """
-    _dom = dominator if dominator is not None else _PARETO_DOMINATOR
-
-    n = len(f)
-    nan_mask = np.any(np.isnan(f), axis=1)
-    valid = np.where(~nan_mask)[0]
-
-    ranks = np.full(n, -1, int)
-    fronts: list[list[int]] = []
-
-    if len(valid) > 0:
-        # Feed only direction-untransformed valid rows to the dominator; the
-        # dominator is responsible for applying the direction transform.
-        g = f[valid].astype(float)
-
-        # Build the full dominance matrix for valid individuals.
-        dom = _dom.dominance_matrix(g, direction)
-
-        # Front-peel: iteratively collect individuals with zero dominators.
-        dominated_by_count = dom.sum(axis=0)  # (k,) — how many dominate each row
-        remaining = np.ones(len(valid), dtype=bool)
-        k = 0
-        while remaining.any():
-            # Local indices (within valid) of the current front.
-            front_local = np.where(remaining & (dominated_by_count == 0))[0]
-            # Map back to global indices.
-            front_global = valid[front_local].tolist()
-            ranks[valid[front_local]] = k
-            fronts.append(front_global)
-            # Remove this front and decrement counts for individuals they dominated.
-            remaining[front_local] = False
-            dominated_by_count -= dom[front_local].sum(axis=0)
-            k += 1
-
-    # Push NaN individuals to a final sentinel front (one per NaN row).
-    last_rank = len(fronts)
-    for i in np.where(nan_mask)[0]:
-        ranks[i] = last_rank
-        fronts.append([int(i)])
-
-    return ranks, fronts
-
-
-def dda_non_dominated_sort(
-    f: np.ndarray,
-    direction: np.ndarray | None = None,
-    *,
-    dominator: Dominator | None = None,
-) -> tuple[np.ndarray, list[list[int]]]:
-    """
-    Non-dominated sorting via the Dominance-Degree Approach (DDA-ENS).
-
-    Produces **identical** ``(ranks, fronts)`` to :func:`non_dominated_sort`
-    and satisfies the :class:`NonDominatedSorter` Protocol.  It is intended
-    as a scalable drop-in for large *N* and/or large *M* (M > 100).
-
-    The dominance relation is computed once as an *NxN* boolean matrix
-    (peak memory O(N^2), **independent of M**; no (N, N, M) tensor is ever
-    allocated).  Front assignment then follows the DDA-ENS scheme: solutions
-    are ordered by ascending dominance-in-degree (number of solutions that
-    dominate them), so that a solution's dominators are always processed
-    before the solution itself.  Each solution is assigned to the first
-    existing front whose members do **not** dominate it, creating a new
-    front when necessary.
-
-    NaN handling is identical to :func:`non_dominated_sort`: rows that
-    contain any NaN value are excluded from the dominance computation and
-    appended afterwards as individual sentinel fronts at rank
-    ``len(real_fronts)``.
-
-    Parameters
-    ----------
-    f : np.ndarray
-        Objective matrix.  shape: (n, n_obj)
-    direction : np.ndarray or None
-        Per-objective optimization direction: +1 = maximize, -1 = minimize.
-        ``None`` defaults to minimization for all objectives.
-    dominator : Dominator or None
-        Dominance predicate to use.  ``None`` defaults to
-        ``ParetoDominator`` (standard Pareto dominance).
-
-    Returns
-    -------
-    ranks : np.ndarray
-        Pareto front index for each individual (0 = first/best front).
-        shape: (n,), dtype int.
-    fronts : list[list[int]]
-        ``fronts[i]`` contains the global indices of individuals in front i.
-
-    References
-    ----------
-    .. [1] Zhou, Y., Chen, Z., & Zhang, J. (2017). Ranking Vectors by Means of
-       the Dominance Degree Matrix. IEEE Transactions on Evolutionary
-       Computation, 21(1), 34-51. https://ieeexplore.ieee.org/document/7469397
-    .. [2] DDA-ENS: Dominance Degree Approach based Efficient Non-dominated
-       Sort. IEEE Conference Publication (2020).
-       https://ieeexplore.ieee.org/document/9282978
-    """
-    _dom = dominator if dominator is not None else _PARETO_DOMINATOR
-
-    n = len(f)
-    nan_mask = np.any(np.isnan(f), axis=1)
-    valid = np.where(~nan_mask)[0]
-
-    ranks = np.full(n, -1, int)
-    fronts: list[list[int]] = []
-
-    if len(valid) > 0:
-        g = f[valid].astype(float)
-        k = len(valid)
-
-        # Build dominance matrix once: D[i, j] = True means valid[i] dominates valid[j].
-        # Memory O(N²), no (N, N, M) tensor.
-        dom_mat = _dom.dominance_matrix(g, direction)
-
-        # in_degree[i] = number of solutions that dominate solution i (locally).
-        in_degree = dom_mat.sum(axis=0)  # (k,)
-
-        # Process solutions in ascending in-degree order so that dominators of
-        # a solution are always assigned to a front before the solution itself.
-        # This guarantees each solution can be placed in the correct front in a
-        # single pass.
-        order = np.argsort(in_degree, kind="stable")
-
-        # front_members[r] = list of local indices assigned to front r.
-        front_members: list[list[int]] = []
-
-        local_ranks = np.full(k, -1, int)
-
-        for i in order:
-            assigned = False
-            # Scan fronts in increasing rank order; assign to the first front
-            # where no current member dominates solution i.
-            for r, members in enumerate(front_members):
-                # dom_mat[m, i] is True if member m dominates solution i.
-                dominated = any(dom_mat[m, i] for m in members)
-                if not dominated:
-                    front_members[r].append(i)
-                    local_ranks[i] = r
-                    assigned = True
-                    break
-            if not assigned:
-                # No existing front is safe; open a new one.
-                local_ranks[i] = len(front_members)
-                front_members.append([i])
-
-        # Map local indices back to global indices and build output.
-        for members in front_members:
-            global_members = valid[members].tolist()
-            fronts.append(global_members)
-            ranks[valid[members]] = local_ranks[members]
-
-    # Append NaN individuals as individual sentinel fronts (same contract as
-    # non_dominated_sort: each NaN row gets rank = len(real_fronts)).
-    last_rank = len(fronts)
-    for i in np.where(nan_mask)[0]:
-        ranks[i] = last_rank
-        fronts.append([int(i)])
-
-    return ranks, fronts
-
-
-class NonDominatedSorter(Protocol):
-    """
-    Protocol for non-dominated sorting callables.
-
-    Any callable that accepts an objective matrix ``f``, an optional
-    per-objective direction array, and an optional ``dominator`` keyword
-    argument and returns ``(ranks, fronts)`` satisfies this protocol.
-    The free function ``non_dominated_sort`` is the default implementation.
-
-    Parameters
-    ----------
-    f : np.ndarray
-        Objective matrix.  shape: (n, n_obj)
-    direction : np.ndarray or None
-        Per-objective optimization direction: +1 = maximize, -1 = minimize.
-        ``None`` defaults to minimization for all objectives.
-    dominator : Dominator or None
-        Dominance predicate to use.  ``None`` defaults to
-        ``ParetoDominator``.
-
-    Returns
-    -------
-    ranks : np.ndarray
-        Pareto front index for each individual (0 = first/best front).
-        shape: (n,)
-    fronts : list[list[int]]
-        ``fronts[i]`` contains the local indices of individuals in front i.
-    """
-
-    def __call__(
-        self,
-        f: np.ndarray,
-        direction: np.ndarray | None = None,
-        *,
-        dominator: Dominator | None = None,
-    ) -> tuple[np.ndarray, list[list[int]]]:
-        """Sort individuals into non-dominated fronts."""
-        ...
-
-
-def crowding_distance(f_front: np.ndarray) -> np.ndarray:
-    """
-    Compute crowding distance for a single Pareto front (NSGA-II).
-
-    Boundary solutions (minimum and maximum per objective) are assigned
-    infinite crowding distance.
-
-    Parameters
-    ----------
-    f_front : np.ndarray
-        Objective values of individuals in the front. shape: (n, n_obj)
-
-    Returns
-    -------
-    np.ndarray
-        Crowding distances. shape: (n,)
-    """
-    n, m = f_front.shape
-    cd = np.zeros(n)
-    if n <= 2:
-        cd[:] = np.inf
-        return cd
-    for obj in range(m):
-        order = np.argsort(f_front[:, obj])
-        cd[order[0]] = np.inf
-        cd[order[-1]] = np.inf
-        f_range = f_front[order[-1], obj] - f_front[order[0], obj]
-        if f_range < 1e-12:
-            continue
-        for k in range(1, n - 1):
-            cd[order[k]] += (
-                f_front[order[k + 1], obj] - f_front[order[k - 1], obj]
-            ) / f_range
-    return cd
-
-
-def crowding_distance_all_fronts(
-    f: np.ndarray,
-    fronts: list[list[int]],
-) -> np.ndarray:
-    """
-    Compute crowding distance for all fronts.
-
-    Parameters
-    ----------
-    f : np.ndarray
-        Objective matrix. shape: (n, n_obj)
-    fronts : list[list[int]]
-        Output of non_dominated_sort: fronts[i] = local indices in front i.
-
-    Returns
-    -------
-    np.ndarray
-        Crowding distances for all individuals. shape: (n,)
-    """
-    cd = np.zeros(len(f))
-    for front in fronts:
-        if not front:
-            continue
-        idx = np.array(front)
-        cd[idx] = crowding_distance(f[idx])
-    return cd
-
-
-def spea2_fitness(
-    f: np.ndarray,
-    direction: np.ndarray | None = None,
-    dominator: Dominator | None = None,
-) -> np.ndarray:
-    """
-    Compute SPEA2 fitness values for a set of objective vectors.
-
-    Implements the fitness assignment procedure from Zitzler et al. (2001).
-    Lower fitness is better (this deviates from the library's general
-    higher-is-better score convention — do NOT pass the result directly
-    to comparators that assume higher = better).
-
-    Fitness components:
-
-    - **Strength** ``S(i)``: number of solutions that ``i`` dominates.
-    - **Raw fitness** ``R(i)``: sum of the strengths of all solutions that
-      dominate ``i``.  ``R(i) = 0`` iff ``i`` is non-dominated.
-    - **Density** ``D(i) = 1 / (sigma_i^k + 2)``, where ``sigma_i^k`` is the
-      Euclidean distance (in objective space) from ``i`` to its k-th nearest
-      neighbour. The ``+2`` keeps ``D in (0, 0.5]``.
-    - **Fitness** ``F(i) = R(i) + D(i)``.
-
-    .. note::
-        **Deviation from the paper**: Zitzler et al. (2001) define
-        ``k = floor(√(N + N̄))`` using the combined size of the population
-        and an external archive.  This utility operates on a single set of
-        ``N`` points, so it uses ``k = floor(√N)`` instead, clamped to
-        ``[0, N-1]``.
-
-    Parameters
-    ----------
-    f : np.ndarray
-        Objective matrix.  shape: (N, n_obj).
-    direction : np.ndarray or None
-        Per-objective optimization direction: ``+1`` = maximize,
-        ``-1`` = minimize.  ``None`` defaults to minimization for all
-        objectives.  Passed directly to the dominator (same convention as
-        :class:`ParetoComparator`).
-    dominator : Dominator or None
-        Dominance predicate used to build the dominance matrix.  ``None``
-        defaults to :class:`ParetoDominator` (standard Pareto dominance).
-
-    Returns
-    -------
-    np.ndarray
-        SPEA2 fitness values ``F``, shape ``(N,)``.  **Lower = better.**
-
-    References
-    ----------
-    .. [1] Zitzler, E., Laumanns, M., & Thiele, L. (2001). SPEA2: Improving
-       the Strength Pareto Evolutionary Algorithm.  TIK-Report 103, ETH
-       Zurich, Switzerland.
-    """
-    f = np.asarray(f, dtype=float)
-    n_pts = len(f)
-
-    if n_pts == 0:
-        return np.empty(0, dtype=float)
-
-    _dom = dominator if dominator is not None else ParetoDominator()
-
-    # --- Strength and raw fitness ---
-    # dom_mat[i, j] = True iff i dominates j
-    dom_mat = _dom.dominance_matrix(f, direction)
-
-    # strength[i] = number of solutions that i dominates (row sum)
-    strength = dom_mat.sum(axis=1).astype(float)
-
-    # raw_fitness[i] = sum of strength[j] for all j that dominate i
-    # dom_mat[j, i] = True iff j dominates i; weight each by strength[j]
-    raw_fitness = dom_mat.T.astype(float) @ strength  # shape (n_pts,)
-
-    # --- Density ---
-    # Pairwise Euclidean distances in objective space; shape (n_pts, n_pts)
-    diff = f[:, None, :] - f[None, :, :]  # (n_pts, n_pts, n_obj)
-    dist_mat = np.sqrt((diff**2).sum(axis=2))  # (n_pts, n_pts)
-
-    # k-th nearest neighbour distance (self-distance 0 sits at index 0)
-    k = int(np.floor(np.sqrt(n_pts)))
-    k = min(k, n_pts - 1)  # clamp to valid index
-
-    # Sort each row ascending and take the k-th column
-    dist_sorted = np.sort(dist_mat, axis=1)  # (n_pts, n_pts)
-    sigma_k = dist_sorted[:, k]  # (n_pts,)
-
-    density = 1.0 / (sigma_k + 2.0)
-
-    return raw_fitness + density
 
 
 class ParetoComparator(Comparator):
@@ -2186,3 +1449,22 @@ class RNSGA2Comparator(ParetoComparator):
                     best_dist[ref] = d_val
 
         return np.array(selected + cleared, dtype=int)
+
+
+# Backward-compatibility re-exports
+from saealib.dominance import (  # noqa: F401
+    Dominator,
+    EpsilonDominator,
+    ParetoDominator,
+    _PARETO_DOMINATOR,
+    _dominance_matrix,
+    _pareto_dominates,
+)
+from saealib.nds import (  # noqa: F401
+    NonDominatedSorter,
+    crowding_distance,
+    crowding_distance_all_fronts,
+    dda_non_dominated_sort,
+    non_dominated_sort,
+    spea2_fitness,
+)
